requirements.txt
flask
flask-cors
python-dotenv
pandas
langchain
deepseek
requests